refactor: move methods to helper class
