+++ app/src/main/java/xtr/keymapper/service/InputListenerService.java
@@ -3,25 +3,25 @@ package xtr.keymapper.service;
 import android.inputmethodservice.InputMethodService;
 import android.view.inputmethod.EditorInfo;
 
-import xtr.keymapper.server.RemoteService;
+import xtr.keymapper.server.RemoteServiceHelper;
 
 public class InputListenerService extends InputMethodService {
     @Override
     public void onDestroy() {
         super.onDestroy();
-        RemoteService.resumeKeymap();
+        RemoteServiceHelper.resumeKeymap();
     }
 
     @Override
     public void onFinishInputView(boolean finishingInput) {
         super.onFinishInputView(finishingInput);
-        RemoteService.resumeKeymap();
+        RemoteServiceHelper.resumeKeymap();
     }
 
     @Override
     public void onStartInputView(EditorInfo info, boolean restarting) {
         super.onStartInputView(info, restarting);
-        RemoteService.pauseKeymap();
+        RemoteServiceHelper.pauseKeymap();
     }
 
 }
